adds colon and elif in new level 17
