--- archive/level20-Additions.lark
@@ -1,3 +1,0 @@
-// adds list access with numbers
-
-?atom: NUMBER | var | NAME | list_access | unsupported_number
