--- archive/old-13-adds-rect-brack.lark
@@ -1,16 +1,0 @@
-// Adds square brackets around lists
-// string list elements also need single quotation marks (') now
-
-// and list items can be assigned to
-
-command: print | ifs (elifs)* elses? | input | for_loop | assign_list | list_access_var | change_list_item | assign
-
-assign: var _SPACE _IS _SPACE sum | var _SPACE _IS _SPACE textwithoutspaces | list_access _SPACE _IS _SPACE sum | list_access _SPACE _IS _SPACE textwithoutspaces
-assign_list: var _SPACE _IS _SPACE _LEFT_SQUARE_BRACKET quoted_text (_COMMA _SPACE? quoted_text)+ _RIGHT_SQUARE_BRACKET
-
-//new for level 12, assignment of list[i]
-change_list_item: var _LEFT_SQUARE_BRACKET (index | var) _RIGHT_SQUARE_BRACKET _SPACE _IS _SPACE (var | textwithoutspaces)
-
-list_access: var _LEFT_SQUARE_BRACKET (index | random | var) _RIGHT_SQUARE_BRACKET
-list_access_var: var _SPACE _IS _SPACE var _LEFT_SQUARE_BRACKET (index | random | var) _RIGHT_SQUARE_BRACKET
-equality_check: (textwithoutspaces | list_access) _SPACE _IS _SPACE (textwithoutspaces | list_access)
