+++ grammars/level10-Additions.lark
@@ -1,5 +1,6 @@
 command: print | repeat | repeat_list | ifs elses? | turtle | ask | assign_list | sleep | list_access_var | assign | comment
 
+//TODO: Maybe this should be called for_list instead of repeat_list? I find thi a bit confusing myself (FH, Nov 2021)
 repeat_list: _FOR _SPACE var _SPACE _IN _SPACE var _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
