+++ grammars/level15-Additions.lark
@@ -1,6 +1,8 @@
 // adds the while loop
 
-command: print | ifs elses? | ask | for_loop | while_loop| assign_list | sleep | list_access_var | assign | comment
+command: print | ifs elses? | ask | for_loop | repeat_list | while_loop| assign_list | sleep | list_access_var | assign | comment
+
+
 while_loop: _WHILE _SPACE (condition) _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
